Statistical inference course project | Part 1
The exponential distribution can be simulated in R with rexp(n, lambda)
where lambda is the rate parameter. The mean of exponential distribution
is 1/lambda and the standard deviation is also also 1/lambda. Set lambda
= 0.2 for all of the simulations. In this simulation, you will
investigate the distribution of averages of 40 exponential(0.2)s. Note
that you will need to do a thousand or so simulated averages of 40
exponentials.